+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 Notable changes to this project.
 
+## [dev]
+- added `download_submission` to download a CSV of a previous submission
+- added `submission_ids` to fetch all submission_ids of a model
+
 ## [2.15.1] - 2023-08-25 - model upload improvements
 - model_upload now allows specifying which data_version and which image to use
 - added `round_model_performances_v2`, which allows fetching newer metrics